### 2019-04-15 d3cb417
Delete paper.bib
