+++ paper.bib
@@ -1,56 +1,0 @@
-@article{10.1186/gb-2008-9-9-r137,
-	author = {Zhang, Yong AND Liu, Tao AND Meyer, Clifford A. AND Eeckhoute, J{\'e}r{\^o}me AND Johnson, David S. AND Bernstein, Bradley E. AND Nusbaum,
-	Chad AND Myers, Richard M. AND Brown, Myles AND Li, Wei AND Liu, X. Shirley},
-	title = {Model-based Analysis of ChIP-Seq (MACS)},
-	journal = {Genome Biology},
-	year = {2008}, month = {09}, volume = {9}, number = {9}, pages = {R137},
-	issn = {1474-760X}, doi = {10.1186/gb-2008-9-9-r137},
-}
-
-@article{10.1371/journal.pone.0065598,
-	author = {Shen, Li AND Shao, Ning-Yi AND Liu, Xiaochuan AND Maze, Ian AND Feng, Jian AND Nestler, Eric J.},
-    journal = {PLoS ONE},
-    publisher = {Public Library of Science},
-    title = {diffReps: Detecting Differential Chromatin Modification Sites from ChIP-seq Data with Biological Replicates},
-	year = {2013}, month = {06}, volume = {8}, number = {6}, pages = {1-13},
-    doi = {10.1371/journal.pone.0065598}
-}
-
-@article{10.1371/journal.pone.0165581,
-    author = {Lin, Jason AND Hiraoka, Kiriko AND Watanabe, Takayoshi AND Kuo, Tony AND Shinozaki, Yoshinao AND Takatori, Atsushi AND Koshikawa, Nobuko AND Chandran, Anandhakumar AND Otsuki, Joe AND Sugiyama, Hiroshi AND Horton, Paul AND Nagase, Hiroki},
-	journal = {PLoS ONE},
-	publisher = {Public Library of Science},
-	title = {Identification of Binding Targets of a Pyrrole-Imidazole Polyamide KR12 in the LS180 Colorectal Cancer Genome},
-	year = {2016}, month = {10}, volume = {11}, number = {10}, pages = {1-19},
-	doi = {10.1371/journal.pone.0165581}
-}
-
-@article{10.1093/bioinformatics/btp352,
-	author = {Li, Heng AND Handsaker, Bob AND Wysoker, Alec AND Fennell, Tim AND Ruan, Jue AND Homer, Nils AND Marth, Gabor AND Abecasis, Goncalo AND Durbin, Richard AND 1000 Genome Project Data Processing Subgroup},
-	title = {The Sequence Alignment/Map format and SAMtools},
-	journal = {Bioinformatics},
-	year = {2009}, volume = {25}, number = {16}, pages = {2078-2079},
-	doi = {10.1093/bioinformatics/btp352}
-}
-
-@inbook{doi:10.1137/1.9781611970586,
-	author = {Durbin, J.},
-	booktitle = {Distribution Theory for Tests Based on the Sample Distribution Function},
-	chapter = {},
-	doi = {10.1137/1.9781611970586},
-	URL = {https://epubs.siam.org/doi/abs/10.1137/1.9781611970586},
-	eprint = {https://epubs.siam.org/doi/pdf/10.1137/1.9781611970586}
-}
-
-@article{10.2307/2346797,
-	ISSN = {00359254, 14679876},
-	URL = {http://www.jstor.org/stable/2346797},
-	author = {Majumder, K.L. AND Bhattacharjee, G. P.},
-	journal = {Journal of the Royal Statistical Society. Series C (Applied Statistics)},
-	number = {3},
-	pages = {409--411},
-	publisher = {[Wiley, Royal Statistical Society]},
-	title = {Algorithm AS 63: The Incomplete Beta Integral},
-	volume = {22},
-	year = {1973}
-}
