### 2019-04-15 ddde0a4
small fixes
