+++ paper.md
@@ -1,56 +1,0 @@
----
-title: 'CRED: a rapid peak caller for Chem-seq data'
-tags:
-  - polyamides
-  - chemical biology
-  - chem-seq
-authors:
- - name: Jason Lin
-   orcid: 0000-0002-8086-3185
-   affiliation: 1 2
- - name: Tony Kuo
-   orcid: 0000-0003-2358-7919
-   affiliation: 2
- - name: Paul Horton
-   orcid: 0000-0002-0916-7339
-   affiliation: 3 4
- - name: Hiroki Nagase
-   orcid: 0000-0002-3992-5399
-   affiliation: 1
-affiliations:
- - name: Laboratory of Cancer Genetics, Chiba Cancer Center Research Institute, Chuo-ku, Chiba, Japan
-   index: 1
-affiliations:
- - name: Artificial Intelligence Research Center, National Institute of Advanced Industrial Science and Technology (AIST), Koto-ku, Tokyo, Japan
-   index: 2
-affiliations:
- - name: Institute of Medical Informatics, National Cheng Kung University, Tainan, Taiwan
-   index: 3
-affiliations:
- - name: Department of Computer Science and Information Engineering, National Cheng Kung University, Tainan, Taiwan
-   index: 4
-date: 8 April 2019
-bibliography: paper.bib
----
-
-# Background
-Chem-Seq Read Enrichment Discovery (``CRED``) is a rapid peak caller written in C for next-generation sequencing (NGS) data, particularly designed with analyzing affinity-enrichment sequencing experiments with pyrrole-imidazole polyamides. Pyrrole-imidazole (PI) polyamides are synthetic molecules, which have primary sequences composed of N-methylpyrrole and N-methylimidazole subunits. These molecules are highly sequence-specific DNA minor-groove binders, but as a consequence of their relatively short recognition motif and molecular weight, the resultant sequencing data often showed a mixture of broad and narrow peaks atypical of other NGS experiments. This fact likely was due to the smaller binding surface in polyamide-DNA interactions compared to their protein-DNA counterparts. We previously designed and reported a workflow to characterize genomic sites enriched with PI polyamide-bound DNA fragments, but the approach required extended preprocessing to convert aligned reads to BED files; additionally, the performance and post-processing of the output unnecessarily lengthened the workflow and hindered throughput. Such shortcomings necessitated computational improvements that remain unmet in the field of Chem-seq research.
-
-# Overview
-Despite that various peak callers for analyzing NGS data exist, most embrace the philosophy that peaks could be modeled as Poisson scattering events; while this approach works well in most ChIP-seq cases, various publications have reported deficiencies of this algorithm in applications such as DNase-seq.
-
-In our research of pyrrole-imidazole polyamides, a class of DNA minor-groove binders, we have also noticed that MACS [Zhang et al. 2008], a popular model-based peak caller, suffered from similar issues in the analysis of affinity-enriched DNA fragments sequenced by Ion Torrent systems (a method we hereafter will refer as "Chem-seq"). We previously proposed a coverage-based approach [Lin et al. 2016], in which we employed Perl-based diffReps [Shen et al. 2013] as the initial candidate selection component, followed by bootstrapped Kolmogorov-Smirnov comparisons to characterize Chem-seq peaks.
-
-This workflow, however, required extensive pre- and post-processing of data, as Perl lacked a direct library to access sequencing data stored in BAM files, a format standard shared by various short-read aligners and NGS tools. While the popular BioPerl package included APIs for processing BAM files, the large amount of dependencies (most of which being unrelated to the workflow) and the size of the BioPerl library made it an undesirable for implementation. There were also performance issues associated with external system calls and the reliance on R to perform more complex mathematical computations.
-
[…]
