### 2019-04-15 da4ff01
more small fixes
