+++ paper/paper.md
@@ -13,7 +13,7 @@ authors:
    affiliation: "2"
  - name: Paul Horton
    orcid: 0000-0002-0916-7339
-   affiliation: "3"
+   affiliation: "3, 4"
  - name: Hiroki Nagase
    orcid: 0000-0002-3992-5399
    affiliation: "1"
@@ -22,14 +22,16 @@ affiliations:
    index: 1
  - name: Artificial Intelligence Research Center, National Institute of Advanced Industrial Science and Technology (AIST), Koto-ku, Tokyo, Japan
    index: 2
- - name: Institute of Medical Informatics and the Department of Computer Science and Information Engineering, National Cheng Kung University, Tainan, Taiwan
+ - name: Institute of Medical Informatics, National Cheng Kung University, Tainan, Taiwan
    index: 3
+ - name: Department of Computer Science and Information Engineering, National Cheng Kung University, Tainan, Taiwan
+   index: 4
 date: 10 April 2019
 bibliography: paper.bib
 ---
 
 # Background
-Chem-Seq Read Enrichment Discovery (``CRED``) is a rapid peak caller written in C for next-generation sequencing (NGS) data, particularly designed with analyzing affinity-enrichment sequencing experiments with pyrrole-imidazole polyamides. Pyrrole-imidazole (PI) polyamides are synthetic molecules, which have primary sequences composed of N-methylpyrrole and N-methylimidazole subunits. These molecules are highly sequence-specific DNA minor-groove binders, but as a consequence of their relatively short recognition motif and molecular weight, the resultant sequencing data often showed a mixture of broad and narrow peaks atypical of other NGS experiments. This fact likely was due to the smaller binding surface in polyamide-DNA interactions compared to their protein-DNA counterparts. We previously designed and reported a workflow to characterize genomic sites enriched with PI polyamide-bound DNA fragments, but the approach required extended preprocessing to convert aligned reads to BED files; additionally, the performance and post-processing of the output unnecessarily lengthened the workflow and hindered throughput. Such shortcomings necessitated computational improvements that remain unmet in the field of Chem-seq research.
+Chem-Seq Read Enrichment Discovery (CRED) is a rapid peak caller written in C for next-generation sequencing (NGS) data, particularly designed with analyzing affinity-enrichment sequencing experiments with pyrrole-imidazole polyamides. Pyrrole-imidazole (PI) polyamides are synthetic molecules, which have primary sequences composed of N-methylpyrrole and N-methylimidazole subunits. These molecules are highly sequence-specific DNA minor-groove binders, but as a consequence of their relatively short recognition motif and molecular weight, the resultant sequencing data often showed a mixture of broad and narrow peaks atypical of other NGS experiments. This fact likely was due to the smaller binding surface in polyamide-DNA interactions compared to their protein-DNA counterparts. We previously designed and reported a workflow to characterize genomic sites enriched with PI polyamide-bound DNA fragments, but the approach required extended preprocessing to convert aligned reads to BED files; additionally, the performance and post-processing of the output unnecessarily lengthened the workflow and hindered throughput. Such shortcomings necessitated computational improvements that remain unmet in the field of Chem-seq research.
 
 # Overview
 Despite that various peak callers for analyzing NGS data exist, most embrace the philosophy that peaks could be modeled as Poisson scattering events; while this approach works well in most ChIP-seq cases, various publications have reported deficiencies of this algorithm in applications such as DNase-seq.
@@ -38,12 +40,12 @@ In our research of pyrrole-imidazole polyamides, a class of DNA minor-groove bin
 
[…]
