### 2019-04-15 0a0717f
more small fixes
