### 2019-04-15 #1374 @jlincbio
Thanks for picking up the review. I noticed that whedon was unable to generate a PDF, so looked around and fixed some things in the paper. Hopefully this wouldn't affect the review process.
