### 2019-04-15 #1374 @jlincbio
@labarba @lpantano this is my first time submitting to JOSS (and TBH first time submitting a software paper), so I'm actually not sure on what sort of criteria a suitable reviewer should meet. With that said, looking at the previous JOSS referee list and perhaps one of the following may work?

MDebasish
darogan
arvestad
janpb
standage
jdeligt

I just filtered the list for the following criteria:
1. Having C/C++ and Perl as the preferred languages
2. Current research area involves some aspect of biology
3. No active reviews at the moment
