### 2019-04-18 #1374 @jlincbio
Thanks for agreeing to handle this submission, @lpantano - it is my first software submission, so I am a bit over-anxious about the whole experience. I made a list of possible referees in an earlier post, but they were not tagged with mentions. If there is anything you need from me please do not hesitate to let me know.
