### 2019-05-01 #1423 @whedon
[REVIEW]: CRED: a rapid peak caller for Chem-seq data
**Submitting author:** @jlincbio (<a href="http://orcid.org/0000-0002-8086-3185">Jason Lin</a>)
**Repository:** <a href="https://github.com/jlincbio/cred" target ="_blank">https://github.com/jlincbio/cred</a>
**Version:** 0.1
