### 2019-05-02 #1423 @darogan
@jlincbio The references don't appear in the article proof - is this some thing you can fix?

Also, would it be appropriate to cite your recent PLoS One (10.1371/journal.pone.0215247) paper as I can imagine both tools being used during a project
