+++ paper/paper.bib
@@ -1,42 +1,57 @@
 @article{10.1186/gb-2008-9-9-r137,
-	author = {Zhang, Yong AND Liu, Tao AND Meyer, Clifford A. AND Eeckhoute, J{\'e}r{\^o}me AND Johnson, David S. AND Bernstein, Bradley E. AND Nusbaum,
-	Chad AND Myers, Richard M. AND Brown, Myles AND Li, Wei AND Liu, X. Shirley},
+	author = {Zhang, Yong and Liu, Tao and Meyer, Clifford A. and Eeckhoute, Jerome and Johnson, David S. and Bernstein, Bradley E. and Nusbaum,
+	Chad and Myers, Richard M. and Brown, Myles and Li, Wei and Liu, X. Shirley},
 	title = {Model-based Analysis of ChIP-Seq (MACS)},
 	journal = {Genome Biology},
-	year = {2008}, month = {09}, volume = {9}, number = {9}, pages = {R137},
-	issn = {1474-760X}, doi = {10.1186/gb-2008-9-9-r137},
+	year = {2008},
+	month = {9},
+	volume = {9},
+	number = {9},
+	pages = {R137},
+	issn = {1474-760X},
+	doi = {10.1186/gb-2008-9-9-r137}
 }
 
 @article{10.1371/journal.pone.0065598,
-	author = {Shen, Li AND Shao, Ning-Yi AND Liu, Xiaochuan AND Maze, Ian AND Feng, Jian AND Nestler, Eric J.},
+	author = {Shen, Li and Shao, Ning-Yi and Liu, Xiaochuan and Maze, Ian and Feng, Jian and Nestler, Eric J.},
     journal = {PLoS ONE},
     publisher = {Public Library of Science},
     title = {diffReps: Detecting Differential Chromatin Modification Sites from ChIP-seq Data with Biological Replicates},
-	year = {2013}, month = {06}, volume = {8}, number = {6}, pages = {1-13},
+	year = {2013},
+	month = {6},
+	volume = {8},
+	number = {6},
+	pages = {1-13},
     doi = {10.1371/journal.pone.0065598}
 }
 
 @article{10.1371/journal.pone.0165581,
-    author = {Lin, Jason AND Hiraoka, Kiriko AND Watanabe, Takayoshi AND Kuo, Tony AND Shinozaki, Yoshinao AND Takatori, Atsushi AND Koshikawa, Nobuko AND Chandran, Anandhakumar AND Otsuki, Joe AND Sugiyama, Hiroshi AND Horton, Paul AND Nagase, Hiroki},
+    author = {Lin, Jason and Hiraoka, Kiriko and Watanabe, Takayoshi and Kuo, Tony and Shinozaki, Yoshinao and Takatori, Atsushi and Koshikawa, Nobuko and Chandran, Anandhakumar and Otsuki, Joe and Sugiyama, Hiroshi and Horton, Paul and Nagase, Hiroki},
 	journal = {PLoS ONE},
 	publisher = {Public Library of Science},
 	title = {Identification of Binding Targets of a Pyrrole-Imidazole Polyamide KR12 in the LS180 Colorectal Cancer Genome},
-	year = {2016}, month = {10}, volume = {11}, number = {10}, pages = {1-19},
+	year = {2016},
+	month = {10},
+	volume = {11},
+	number = {10},
+	pages = {1-19},
 	doi = {10.1371/journal.pone.0165581}
 }
 
 @article{10.1093/bioinformatics/btp352,
[…]
