+++ paper/paper.bib
@@ -39,7 +39,7 @@
 }
 
 @article{10.1093/bioinformatics/btp352,
-	author = {Li, Heng and Handsaker, Bob and Wysoker, Alec and Fennell, Tim and Ruan, Jue and Homer, Nils and Marth, Gabor and Abecasis, Goncalo and Durbin, Richard and 1000 Genome Project Data Processing Subgroup},
+	author = {Li, H. and Handsaker, B. and Wysoker, A. and Fennell, T. and Ruan, J. and Homer, N.i and Marth, G. and Abecasis, G. and Durbin, R. and {1000 Genome Project Data Processing Subgroup}},
 	title = {The Sequence Alignment/Map format and SAMtools},
 	journal = {Bioinformatics},
 	year = {2009},
@@ -49,24 +49,23 @@
 	doi = {10.1093/bioinformatics/btp352}
 }
 
-@inbook{doi:10.1137/1.9781611970586,
-	author = {Durbin, J.},
-	booktitle = {Distribution Theory for Tests Based on the Sample Distribution Function},
-	doi = {10.1137/1.9781611970586},
-	URL = {https://epubs.siam.org/doi/abs/10.1137/1.9781611970586},
-	eprint = {https://epubs.siam.org/doi/pdf/10.1137/1.9781611970586}
-}
-
 @article{10.2307/2346797,
 	ISSN = {00359254, 14679876},
 	URL = {http://www.jstor.org/stable/2346797},
 	author = {Majumder, K.L. and Bhattacharjee, G. P.},
-	journal = {Journal of the Royal Statistical Society. Series C (Applied Statistics)},
+	journal = {Journal of the Royal Statistical Society Series C (Applied Statistics)},
 	number = {3},
 	pages = {409-411},
-	publisher = {[Wiley, Royal Statistical Society]},
+	publisher = {{Wiley, Royal Statistical Society}},
 	title = {Algorithm AS 63: The Incomplete Beta Integral},
 	volume = {22},
 	year = {1973},
 	doi = {10.2307/2346797}
+}
+
+@inbook{10.1137/1.9781611970586,
+	author = {Durbin, J.},
+	booktitle = {Distribution Theory for Tests Based on the Sample Distribution Function},
+	doi = {10.1137/1.9781611970586},
+	URL = {https://epubs.siam.org/doi/abs/10.1137/1.9781611970586},
 }
