+++ paper/paper.bib
@@ -67,5 +67,4 @@
 	author = {Durbin, J.},
 	booktitle = {Distribution Theory for Tests Based on the Sample Distribution Function},
 	doi = {10.1137/1.9781611970586},
-	URL = {https://epubs.siam.org/doi/abs/10.1137/1.9781611970586},
 }
