### 2019-05-02 #1423 @jlincbio
@darogan thanks for reviewing CRED, Dr. Hamilton. I apologize for the delay.

I am trying to figure out why the references are not showing up - I suspected that it is the bibTex entries but paper.bib seemed to pass the BibTex quality checker (https://biblatex-linter.herokuapp.com/validate); I will continue to try to fix it.

Re: your comment about the recent PLoS ONE publication -  I did not explicitly refer to it in the manuscript and because that publication dealt more with microarray data rather than sequencing data, I decided to leave it as is. Nevertheless I will take up your suggestion and see if it's worth adding a remark in the paper and including this citation.
