+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@article{10.1186/gb-2008-9-9-r137,
+@article{Zhang2008,
 	author = {Zhang, Yong and Liu, Tao and Meyer, Clifford A. and Eeckhoute, Jerome and Johnson, David S. and Bernstein, Bradley E. and Nusbaum,
 	Chad and Myers, Richard M. and Brown, Myles and Li, Wei and Liu, X. Shirley},
 	title = {Model-based Analysis of ChIP-Seq (MACS)},
@@ -12,7 +12,7 @@
 	doi = {10.1186/gb-2008-9-9-r137}
 }
 
-@article{10.1371/journal.pone.0065598,
+@article{Shen2013,
 	author = {Shen, Li and Shao, Ning-Yi and Liu, Xiaochuan and Maze, Ian and Feng, Jian and Nestler, Eric J.},
     journal = {PLoS ONE},
     publisher = {Public Library of Science},
@@ -25,7 +25,7 @@
     doi = {10.1371/journal.pone.0065598}
 }
 
-@article{10.1371/journal.pone.0165581,
+@article{Lin2016,
     author = {Lin, Jason and Hiraoka, Kiriko and Watanabe, Takayoshi and Kuo, Tony and Shinozaki, Yoshinao and Takatori, Atsushi and Koshikawa, Nobuko and Chandran, Anandhakumar and Otsuki, Joe and Sugiyama, Hiroshi and Horton, Paul and Nagase, Hiroki},
 	journal = {PLoS ONE},
 	publisher = {Public Library of Science},
@@ -38,7 +38,7 @@
 	doi = {10.1371/journal.pone.0165581}
 }
 
-@article{10.1093/bioinformatics/btp352,
+@article{Li2009,
 	author = {Li, H. and Handsaker, B. and Wysoker, A. and Fennell, T. and Ruan, J. and Homer, N.i and Marth, G. and Abecasis, G. and Durbin, R. and {1000 Genome Project Data Processing Subgroup}},
 	title = {The Sequence Alignment/Map format and SAMtools},
 	journal = {Bioinformatics},
@@ -49,7 +49,7 @@
 	doi = {10.1093/bioinformatics/btp352}
 }
 
-@article{10.2307/2346797,
+@article{Majumder1973,
 	ISSN = {00359254, 14679876},
 	URL = {http://www.jstor.org/stable/2346797},
 	author = {Majumder, K.L. and Bhattacharjee, G. P.},
@@ -63,7 +63,7 @@
 	doi = {10.2307/2346797}
 }
 
-@inbook{10.1137/1.9781611970586,
+@inbook{Durbin1973,
 	author = {Durbin, J.},
 	booktitle = {Distribution Theory for Tests Based on the Sample Distribution Function},
 	doi = {10.1137/1.9781611970586},
