### 2019-05-03 f8c9a84
Delete paper.bib
