+++ paper.bib
@@ -1,71 +1,0 @@
-@article{Zhang2008,
-	author = {Zhang, Yong and Liu, Tao and Meyer, Clifford A. and Eeckhoute, Jerome and Johnson, David S. and Bernstein, Bradley E. and Nusbaum,
-	Chad and Myers, Richard M. and Brown, Myles and Li, Wei and Liu, X. Shirley},
-	title = {Model-based Analysis of ChIP-Seq (MACS)},
-	journal = {Genome Biology},
-	year = {2008},
-	month = {9},
-	volume = {9},
-	number = {9},
-	pages = {R137},
-	issn = {1474-760X},
-	doi = {10.1186/gb-2008-9-9-r137}
-}
-
-@article{Shen2013,
-	author = {Shen, Li and Shao, Ning-Yi and Liu, Xiaochuan and Maze, Ian and Feng, Jian and Nestler, Eric J.},
-    journal = {PLoS ONE},
-    publisher = {Public Library of Science},
-    title = {diffReps: Detecting Differential Chromatin Modification Sites from ChIP-seq Data with Biological Replicates},
-	year = {2013},
-	month = {6},
-	volume = {8},
-	number = {6},
-	pages = {1-13},
-    doi = {10.1371/journal.pone.0065598}
-}
-
-@article{Lin2016,
-    author = {Lin, Jason and Hiraoka, Kiriko and Watanabe, Takayoshi and Kuo, Tony and Shinozaki, Yoshinao and Takatori, Atsushi and Koshikawa, Nobuko and Chandran, Anandhakumar and Otsuki, Joe and Sugiyama, Hiroshi and Horton, Paul and Nagase, Hiroki},
-	journal = {PLoS ONE},
-	publisher = {Public Library of Science},
-	title = {Identification of Binding Targets of a Pyrrole-Imidazole Polyamide KR12 in the LS180 Colorectal Cancer Genome},
-	year = {2016},
-	month = {10},
-	volume = {11},
-	number = {10},
-	pages = {1-19},
-	doi = {10.1371/journal.pone.0165581}
-}
-
-@article{Li2009,
-	author = {Li, H. and Handsaker, B. and Wysoker, A. and Fennell, T. and Ruan, J. and Homer, N.i and Marth, G. and Abecasis, G. and Durbin, R. and {1000 Genome Project Data Processing Subgroup}},
-	title = {The Sequence Alignment/Map format and SAMtools},
-	journal = {Bioinformatics},
-	year = {2009},
-	volume = {25},
-	number = {16},
-	pages = {2078-2079},
-	doi = {10.1093/bioinformatics/btp352}
-}
-
-@article{Majumder1973,
-	ISSN = {00359254, 14679876},
-	URL = {http://www.jstor.org/stable/2346797},
-	author = {Majumder, K.L. and Bhattacharjee, G. P.},
-	journal = {Journal of the Royal Statistical Society Series C (Applied Statistics)},
-	number = {3},
-	pages = {409-411},
-	publisher = {{Wiley, Royal Statistical Society}},
-	title = {Algorithm AS 63: The Incomplete Beta Integral},
-	volume = {22},
-	year = {1973},
-	doi = {10.2307/2346797}
-}
-
-@inbook{Durbin1973,
-	author = {Durbin, J.},
-	booktitle = {Distribution Theory for Tests Based on the Sample Distribution Function},
-	year = {1973},
-	doi = {10.1137/1.9781611970586},
-}
