+++ paper/paper.bib
@@ -66,5 +66,6 @@
 @inbook{Durbin1973,
 	author = {Durbin, J.},
 	booktitle = {Distribution Theory for Tests Based on the Sample Distribution Function},
+	year = {1973},
 	doi = {10.1137/1.9781611970586},
 }
