### 2019-05-03 #1423 @jlincbio
@darogan sorry about the delay. It turned out that I didn't properly set the reference property names properly and was not calling them in the manuscript. This latest version seemed to have fixed it:

>
>
> [ 👉 Check article proof 📄 👈 ](https://github.com/openjournals/joss-papers/blob/joss.01423/joss.01423/10.21105.joss.01423.pdf)
