### 2019-05-03 #1423 @darogan
@jlincbio Thanks for fixing it, PDF looks great, all references OK.

@lpantano I'm happy with the software/paper - shall I close the issue to complete the review?
