### 2019-05-03 #1423 @lpantano
@jlincbio , we are almost there. Can you create a zenodo doi (https://zenodo.org/) pointed to the current version of the repository? The title of that archive should match paper title and the authors should match as well. Here is an example of another paper:

https://zenodo.org/record/2645762#.XMyf2akpCpc

https://github.com/openjournals/joss-reviews/issues/1364

Cheers
