### 2019-05-04 #1423 @kyleniemeyer
Hi @jlincbio, overall looks good, but before we fully accept I have two comments on the paper. Could you please:
[ ] Add to the background section a bit more helping describe the high-level functionality and purpose of the software for a *diverse, non-specialist audience*? This is part of our expectations for articles, since the JOSS audience is quite broad.
[ ] Add some metadata to the first reference (Durbin 1973)? It looks like it may be a textbook, but if so is missing some info in the reference.
