### 2019-05-04 #1423 @jlincbio
@kyleniemeyer thanks for the comments, I'll try to get right onto it. Re: point 1, should some general description on what a peak caller does be sufficient?

Also, do I need to submit a new release and get a new zenodo DOI?
