+++ paper/paper.bib
@@ -1,6 +1,30 @@
+@article{Dervan2003,
+	title = {Recognition of the DNA minor groove by pyrrole-imidazole polyamides},
+	journal = {Current Opinion in Structural Biology},
+	volume = {13},
+	number = {3},
+	pages = {284-299},
+	year = {2003},
+	issn = {0959-440X},
+	doi = {https://doi.org/10.1016/S0959-440X(03)00081-2},
+	url = {http://www.sciencedirect.com/science/article/pii/S0959440X03000812},
+	author = {Dervan, P.B. and Edelson, B.S.}
+}
+
+@article{Pandian2014,
+	author = {Pandian, G.N. and Sato, S. and Anandhakumar, C. and Taniguchi, J. and Takashima, K. and Syed, J. and Han, L. and Saha, A. and Bando, T. and Nagase, H. and Sugiyama, H.},
+	title = {Identification of a Small Molecule That Turns ON the Pluripotency Gene Circuitry in Human Fibroblasts},
+	journal = {ACS Chemical Biology},
+	volume = {9},
+	number = {12},
+	pages = {2729-2736},
+	year = {2014},
+	doi = {10.1021/cb500724t},
+	url = {https://doi.org/10.1021/cb500724t}
+}
+
 @article{Zhang2008,
-	author = {Zhang, Yong and Liu, Tao and Meyer, Clifford A. and Eeckhoute, Jerome and Johnson, David S. and Bernstein, Bradley E. and Nusbaum,
-	Chad and Myers, Richard M. and Brown, Myles and Li, Wei and Liu, X. Shirley},
+	author = {Zhang, Y. and Liu, T. and Meyer, CA. and Eeckhoute, J. and Johnson, D.S. and Bernstein, B.E. and Nusbaum, C. and Myers, R.M. and Brown, M. and Li, W and Liu, X.S.},
 	title = {Model-based Analysis of ChIP-Seq (MACS)},
 	journal = {Genome Biology},
 	year = {2008},
@@ -13,7 +37,7 @@
 }
 
 @article{Shen2013,
-	author = {Shen, Li and Shao, Ning-Yi and Liu, Xiaochuan and Maze, Ian and Feng, Jian and Nestler, Eric J.},
+	author = {Shen, L. and Shao, N.-Y. and Liu, X. and Maze, I. and Feng, J. and Nestler, E.J.},
     journal = {PLoS ONE},
     publisher = {Public Library of Science},
     title = {diffReps: Detecting Differential Chromatin Modification Sites from ChIP-seq Data with Biological Replicates},
@@ -26,7 +50,7 @@
 }
 
 @article{Lin2016,
-    author = {Lin, Jason and Hiraoka, Kiriko and Watanabe, Takayoshi and Kuo, Tony and Shinozaki, Yoshinao and Takatori, Atsushi and Koshikawa, Nobuko and Chandran, Anandhakumar and Otsuki, Joe and Sugiyama, Hiroshi and Horton, Paul and Nagase, Hiroki},
+    author = {Lin, J. and Hiraoka, K. and Watanabe, T. and Kuo, T. and Shinozaki, Y. and Takatori, A. and Koshikawa, N. and Chandran, A. and Otsuki, J. and Sugiyama, H. and Horton, P. and Nagase, H.},
 	journal = {PLoS ONE},
 	publisher = {Public Library of Science},
 	title = {Identification of Binding Targets of a Pyrrole-Imidazole Polyamide KR12 in the LS180 Colorectal Cancer Genome},
@@ -52,7 +76,7 @@
 @article{Majumder1973,
 	ISSN = {00359254, 14679876},
[…]
