### 2019-05-04 #1423 @jlincbio
> [ 👉 Check article proof 📄 👈 ](https://github.com/openjournals/joss-papers/blob/joss.01423/joss.01423/10.21105.joss.01423.pdf)

@kyleniemeyer would the above changes be sufficient to satisfy the journal requirements?

Also if a new Zenodo archive DOI is necessary I think this is it:
10.5281/zenodo.2667425
