+++ paper/paper.bib
@@ -6,7 +6,7 @@
 	pages = {284-299},
 	year = {2003},
 	issn = {0959-440X},
-	doi = {https://doi.org/10.1016/S0959-440X(03)00081-2},
+	doi = {10.1016/S0959-440X(03)00081-2},
 	url = {http://www.sciencedirect.com/science/article/pii/S0959440X03000812},
 	author = {Dervan, P.B. and Edelson, B.S.}
 }
