+++ paper/paper.md
@@ -31,7 +31,7 @@ bibliography: paper.bib
 ---
 
 # Background
-Chem-Seq Read Enrichment Discovery (CRED) is a rapid peak caller written in C for next-generation sequencing (NGS) data, particularly designed with analyzing affinity-enrichment sequencing experiments with pyrrole-imidazole polyamides. Pyrrole-imidazole (PI) polyamides are synthetic molecules, which have primary sequences composed of N-methylpyrrole and N-methylimidazole subunits with highly sequence-specific DNA minor-groove binding [@Dervan2003]. Upon functionalization with other small molecules such as alkylating agents [@Hiraoka2015] or histone deacetylase inhibitors [@Pandian2014], PI polyamides provide a conduit for those small molecules to interact with specific regions of the genome. PI polyamides' relatively short recognition motif and molecular weight, however, can result in generally smaller binding surfaces in polyamide-DNA interactions compared to interactions of DNA with other biomolecules, for instance proteins and transcription factors. A direct consequence of this difference in interaction leads to a mixture of broad and narrow peaks in sequencing experiments conducted with PI polyamides (Chem-seq) that can be atypical of other NGS experiments [@Lin2016]. To properly analyze Chem-seq data necessitates the creation of a Chem-seq specific computational tool that can analyze general regions of enriched precipitation of polyamide-DNA ligands (a process known as *peak calling*) from sequencing reads in the genome. We previously designed and reported a workflow to characterize genomic sites enriched with PI polyamide-bound DNA fragments, but the approach required extended preprocessing to convert aligned reads (typically stored in BAM, a compressed binary standard) to BED files, a popular tab-delimited flat text format for storing positional data in the genome. As NGS data can reach upwards of tens to hundreds of gigabytes, this conversion step required a significant amount of time and buffer storage; in addition, the performance and post-processing of the output unnecessarily lengthened the workflow further and hindered throughput. Such shortcomings necessitated computational improvements that remain unmet in the field of Chem-seq research.
[…]
