### 2019-05-05 #1423 @kyleniemeyer
Hi @jlincbio, sorry the Durbin ref was still not showing up correctly; I just submitted a [PR](https://github.com/jlincbio/cred/pull/1) to fix it.
