### 2019-05-05 bc43c92
Update paper.bib
