+++ paper/paper.bib
@@ -102,8 +102,8 @@
 	doi = {10.1137/1.9781611970586},
 	url = {https://doi.org/10.1137/1.9781611970586},
 	year = {1973},
-	month = jan,
+	month = {Jan},
 	publisher = {Society for Industrial and Applied Mathematics},
-	author = {J. Durbin},
+	author = {Durbin, J.},
 	title = {Distribution Theory for Tests Based on the Sample Distribution Function}
 }
