### 2019-05-05 #1423 @jlincbio
@kyleniemeyer thanks for the help. I merged the fix on my mobile phone and re-released it under the same version tag of v0.1...hopefully I did not do something stupid to mess it up.

The DOI seems to be the same as before at 10.5281/zenodo.2667428 (the same commit, ddef12d).

Thanks - it's my first time with Zenodo.
