+++ paper/paper.bib
@@ -103,7 +103,7 @@
 	url = {https://doi.org/10.1137/1.9781611970586},
 	year = {1973},
 	month = {Jan},
-	publisher = {Society for Industrial and Applied Mathematics},
+	publisher = {Society for Industrial \& Applied Mathematics},
 	author = {Durbin, J.},
 	title = {Distribution Theory for Tests Based on the Sample Distribution Function}
 }
