### 2019-05-05 #1423 @kyleniemeyer
@jlincbio ok, you are all set. Unfortunately, we have to pause the final publishing step until CrossRef resolves some issues they are dealing with at the moment. Hopefully those are resolved by tomorrow.
