+++ paper/paper.bib
@@ -1,56 +1,109 @@
-@article{10.1186/gb-2008-9-9-r137,
-	author = {Zhang, Yong AND Liu, Tao AND Meyer, Clifford A. AND Eeckhoute, J{\'e}r{\^o}me AND Johnson, David S. AND Bernstein, Bradley E. AND Nusbaum,
-	Chad AND Myers, Richard M. AND Brown, Myles AND Li, Wei AND Liu, X. Shirley},
-	title = {Model-based Analysis of ChIP-Seq (MACS)},
-	journal = {Genome Biology},
-	year = {2008}, month = {09}, volume = {9}, number = {9}, pages = {R137},
-	issn = {1474-760X}, doi = {10.1186/gb-2008-9-9-r137},
+@article{Dervan2003,
+	title = {Recognition of the DNA minor groove by pyrrole-imidazole polyamides},
+	journal = {Current Opinion in Structural Biology},
+	volume = {13},
+	number = {3},
+	pages = {284-299},
+	year = {2003},
+	issn = {0959-440X},
+	doi = {10.1016/S0959-440X(03)00081-2},
+	url = {http://www.sciencedirect.com/science/article/pii/S0959440X03000812},
+	author = {Dervan, P.B. and Edelson, B.S.}
 }
 
-@article{10.1371/journal.pone.0065598,
-	author = {Shen, Li AND Shao, Ning-Yi AND Liu, Xiaochuan AND Maze, Ian AND Feng, Jian AND Nestler, Eric J.},
+@article{Pandian2014,
+	author = {Pandian, G.N. and Sato, S. and Anandhakumar, C. and Taniguchi, J. and Takashima, K. and Syed, J. and Han, L. and Saha, A. and Bando, T. and Nagase, H. and Sugiyama, H.},
+	title = {Identification of a Small Molecule That Turns ON the Pluripotency Gene Circuitry in Human Fibroblasts},
+	journal = {ACS Chemical Biology},
+	volume = {9},
+	number = {12},
+	pages = {2729-2736},
+	year = {2014},
+	doi = {10.1021/cb500724t},
+	url = {https://doi.org/10.1021/cb500724t}
+}
+
+@article{Zhang2008,
+	author = {Zhang, Y. and Liu, T. and Meyer, CA. and Eeckhoute, J. and Johnson, D.S. and Bernstein, B.E. and Nusbaum, C. and Myers, R.M. and Brown, M. and Li, W and Liu, X.S.},
+	title = {Model-based Analysis of ChIP-Seq (MACS)},
+	journal = {Genome Biology},
+	year = {2008},
+	month = {9},
+	volume = {9},
+	number = {9},
+	pages = {R137},
+	issn = {1474-760X},
+	doi = {10.1186/gb-2008-9-9-r137}
+}
+
+@article{Shen2013,
+	author = {Shen, L. and Shao, N.-Y. and Liu, X. and Maze, I. and Feng, J. and Nestler, E.J.},
     journal = {PLoS ONE},
     publisher = {Public Library of Science},
     title = {diffReps: Detecting Differential Chromatin Modification Sites from ChIP-seq Data with Biological Replicates},
-	year = {2013}, month = {06}, volume = {8}, number = {6}, pages = {1-13},
+	year = {2013},
+	month = {6},
+	volume = {8},
+	number = {6},
+	pages = {1-13},
     doi = {10.1371/journal.pone.0065598}
 }
 
-@article{10.1371/journal.pone.0165581,
[…]
