+++ README.md
@@ -2,7 +2,10 @@
 
 **C**hem-seq **R**ead **E**nrichment **D**iscovery (__CRED__) is a simple peak caller written in C for identifying non-canonical feature enrichments in paired Chem-seq data.
 
-Committed April 10, 2019
+[![status](http://joss.theoj.org/papers/f9d17ffdcd6b02ef2f2a5eaa3638c294/status.svg)](http://joss.theoj.org/papers/f9d17ffdcd6b02ef2f2a5eaa3638c294)
+
+04/10/2019 initial commit/*JOSS* submission<br>
+05/04/2019 *JOSS* peer review completed<br>
 
 ### Installation
 
