### 2019-05-07 #1423 @jlincbio
I noticed a small typo (extra "an" in "...BioPerl library made it an undesireable") and modified `paper.md` by the way; hopefully this isn't too late.
