### 2019-05-08 #1423 @kyleniemeyer
@jlincbio congrats on your paper's publication in JOSS! Thanks to @darogan for reviewing and @lpantano for editing!
