### 2019-05-08 #1423 @jlincbio
@lpantano @darogan @kyleniemeyer thanks again for the help along the way and I do apologize for all the inconvenience I've caused. Looking forward to the next chance working with the wonderful people at JOSS in the future.
